backend/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
pillow
bleach
celery
redis
django-cors-headers
